provide remote endpoint via additional read handler
The existing read handler only receives a data buffer, the sender is unavailable. Instead of using async_receive on the socket, async_receive_from is used and the endpoint stored until the read event handler is called.

Stephen provided another solution that did not require any API changes in #6, but unfortunately I was unable to get that working. I don't particularly like this solution, but until another is found, the show must go on…